pyproject.toml
[project]
name = "minbot"
version = "0.1.0"
description = "Minimal Telegram bot for GitHub development assistance"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=21.0",
    "anthropic>=0.40.0",
    "apscheduler>=3.10.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
